refactor(webapp,rtprvw): Extract method to fit map to endpoints
