+++ lib/components/app/responsive-webapp.js
@@ -26,6 +26,7 @@ import * as mapActions from '../../actions/map'
 import * as uiActions from '../../actions/ui'
 import { AUTH0_AUDIENCE, AUTH0_SCOPE } from '../../util/constants'
 import { ComponentContext } from '../../util/contexts'
+import { fitMapToEndpoints } from '../../util/ui'
 import { getActiveItinerary } from '../../util/state'
 import { getAuth0Config } from '../../util/auth'
 import { getDefaultLocale } from '../../util/i18n'
@@ -113,16 +114,7 @@ class ResponsiveWebapp extends Component {
         query.from !== prevProps.query.from &&
         query.to !== prevProps.query.to
       ) {
-        // Recent versions of maplibre enforce the order of coordinates (i.e., sw, ne).
-        const minlat = Math.min(query.from.lat, query.to.lat)
-        const maxlat = Math.max(query.from.lat, query.to.lat)
-        const minlon = Math.min(query.from.lon, query.to.lon)
-        const maxlon = Math.max(query.from.lon, query.to.lon)
-
-        map?.fitBounds([minlon, minlat, maxlon, maxlat], {
-          duration: 600,
-          padding: util.getFitBoundsPadding(map, 0.2)
-        })
+        fitMapToEndpoints(map, query.from, query.to)
       } else if (query.from && !query.to) {
         setMapCenter(map, query.from)
       } else if (query.to && !query.from) {
