+++ lib/util/ui.ts
@@ -1,5 +1,8 @@
 /* eslint-disable camelcase */
 import { FormikProps } from 'formik'
+import { Location } from '@opentripplanner/types'
+import { MapRef } from 'react-map-gl/maplibre'
+import { util } from '@opentripplanner/base-map'
 
 /**
  * @param string the string to test.
@@ -141,5 +144,27 @@ export function getMapToggleNewItineraryView(
       return ItineraryView.LEG
     default:
       return ItineraryView.LIST
+  }
+}
+
+/**
+ * Fit map bounds so that both the from and to endpoints are visible.
+ */
+export function fitMapToEndpoints(
+  map: MapRef,
+  from: Location,
+  to: Location
+): void {
+  if (map) {
+    // Recent versions of maplibre enforce the order of coordinates (i.e., sw, ne).
+    const minlat = Math.min(from.lat, to.lat)
+    const maxlat = Math.max(from.lat, to.lat)
+    const minlon = Math.min(from.lon, to.lon)
+    const maxlon = Math.max(from.lon, to.lon)
+
+    map.fitBounds([minlon, minlat, maxlon, maxlat], {
+      duration: 600,
+      padding: util.getFitBoundsPadding(map, 0.2)
+    })
   }
 }
